cleaned troubleshooting code and fixed missing trains
started to clear out random prints that were no longer needed.  More importantly, the script was missing trains. Before the number of trains it was looking at was hard coded because if it looped through too many trains it got an index error.  Now it loops through as many times as there are trains.